Move cli_options to app namespace
