+++ spyderlib/app/cli_options.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright © 2012 The Spyder development team
+# Copyright © 2009- The Spyder development team
 # Licensed under the terms of the MIT License
 # (see spyderlib/__init__.py for details)
