+++ spyderlib/app/start.py
@@ -10,7 +10,7 @@ import sys
 import time
 
 # Local imports
-from spyderlib.cli_options import get_options
+from spyderlib.app.cli_options import get_options
 from spyderlib.config.base import get_conf_path, running_in_mac_app, DEV, TEST
 from spyderlib.config.main import CONF
 from spyderlib.utils.external import lockfile
